mover textTohtml del repositorio al un Helper para la vista
